Fix #1 - Item table is now indexed correctly
The issue was caused by the way the mod stored the items. It took the
items from an array and stored them in a table starting at 0. The
problem with this approach is, that Lua's length operator doesn't work
correctly with tables starting at zero. Therefore it always skipped the
first item.