Close #4 - Allow partial unpacking of a bag
Instead of blocking the unpacking of a bag completely, when the capacity
of the target container is too small to fit all items, we now now allow
only a portion of the items in the bag to be unpacked.

Currently this increases the duration of the unpacking to represent
how the player is emptying the bag more carefully in order not to
overfill the container.