Close #7 - Allow player to tag bags for quick sorting of certain items
Added a text box which allows the player to add new tags (e.g.: Food,
Nails, Baseball) and remove tags (e.g.: !Food, !Nails, !Baseball).

These tags will be used when a bag is "auto-packed". If an item's
category or name matches one of the tags it will be transferred into the
bag. If it doesn't it will be ignored.